* [misc] add install sql.

+++ db/zentao.sql
@@ -2032,7 +2032,7 @@ CREATE TABLE `zt_taskteam` (
   `left` decimal(12,2) NOT NULL DEFAULT '0.00',
   `transfer` char(30) NOT NULL DEFAULT '',
   `status` enum('wait','doing','done') NOT NULL DEFAULT 'wait',
-  `order` tinyint(3) NOT NULL DEFAULT '0',
+  `order` int(8) NOT NULL DEFAULT '0',
   PRIMARY KEY (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 CREATE INDEX `task` ON `zt_taskteam` (`task`);
